+++ packages/expo-location/android/src/main/java/expo/modules/location/LocationParams.kt
@@ -1,0 +1,10 @@
+package expo.modules.location
+
+data class LocationParams(val accuracy: LocationAccuracy?, var distance: Float, var interval: Long)
+
+enum class LocationAccuracy {
+  LOWEST,
+  LOW,
+  MEDIUM,
+  HIGH
+}